+++ frontend/hud/src/pond.rs
@@ -1,3 +1,4 @@
+use crate::Tank;
 use failure::Error;
 use yew::callback::Callback;
 use yew::format::{Json, Nothing};
@@ -36,21 +37,4 @@ impl PondService {
         let request = Request::get(url.as_str()).body(Nothing).unwrap();
         self.web.fetch(request, handler.into())
     }
-}
-
-/// A struct to hold data returned by the HTTP request
-/// for tanks' temp & ph info.
-#[derive(Debug, Deserialize)]
-pub struct Tank {
-    pub id: u16,
-    pub name: Option<String>,
-    pub temp_f: Option<f32>,
-    pub temp_c: Option<f32>,
-    pub temp_update_time: Option<u64>,
-    pub temp_update_count: Option<u32>,
-
-    pub ph: Option<f32>,
-    pub ph_mv: Option<f32>,
-    pub ph_update_time: Option<u64>,
-    pub ph_update_count: Option<u32>,
 }
